requirements.txt
pyrogram
tgcrypto
requests
numpy
pytesseract
Pillow
